+++ module/action.sh
@@ -1,5 +1,4 @@
 #!/system/bin/sh
 
 MODDIR=${0%/*}
-exec "$MODDIR/bin/mipushctl" scan
-
+exec "$MODDIR/bin/mipushctl" manage